# dash-backend/rest_api/swagger/index.yaml
definitions:
...
  PriceOptionAssignmentSchema:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      end_date:
        format: date-time
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      modification_date:
        format: date-time
        readOnly: true
        type: string
      region_id:
        format: int32
        type: integer
      start_date:
        format: date-time
        type: string
      value_in_cent:
        format: int32
        type: integer
      vehicle_option_item_id:
        format: int32
        type: integer
    type: object
...
  PriceVehicleModelStyleAssignmentSchema:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      end_date:
        format: date-time
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      modification_date:
        format: date-time
        readOnly: true
        type: string
      region_id:
        format: int32
        type: integer
      start_date:
        format: date-time
        type: string
      value_in_cent:
        format: int32
        type: integer
      vehicle_model_style_id:
        format: int32
        type: integer
    type: object
...
  VehicleAttributeSchema:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      description:
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
      parent:
        allOf:
        - $ref: '#/definitions/VehicleAttributeSchema'
        readOnly: true
      parent_id:
        format: int32
        type: integer
    required:
    - description
    - name
    type: object
...
  VehicleBody:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      description:
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
  VehicleBrand:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
...
  VehicleDriveTrain:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      description:
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
  VehicleEngine:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      horse_power:
        format: int32
        type: integer
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      label:
        type: string
      modification_date:
        format: date-time
        readOnly: true
        type: string
      torque:
        format: int32
        type: integer
      type:
        enum:
        - OTHERS
        - GAS
        - DIESEL
        - ELECTRIC
        type: string
    required:
    - type
    type: object
  VehicleGalleryItemSchema:
    properties:
      asset_path:
        type: string
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      modification_date:
        format: date-time
        readOnly: true
        type: string
      order:
        format: int32
        type: integer
      type:
        enum:
        - COVER
        - INTERIOR
        - INTERIOR_COLOR
        - EXTERIOR
        - EXTERIOR_COLOR
        - BRAND_LOGO
        - MODEL_AVATAR
        - YEAR_MODEL_AVATAR
        - VIDEO_INTRODUCTION
        - MODEL_STYLE_AVATAR
        - OPTION_ITEM_AVATAR
        - VEHICLE_AVATAR
        - OTHERS
        type: string
    required:
    - asset_path
    - type
    type: object
...
  VehicleInfo:
    properties:
      city_mpg:
        format: int32
        type: integer
      combined_mpg:
        format: int32
        type: integer
      creation_date:
        format: date-time
        readOnly: true
        type: string
      description:
        type: string
      doors:
        format: int32
        type: integer
      highway_mpg:
        format: int32
        type: integer
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      safety_rate:
        format: int32
        type: integer
      seats:
        format: int32
        type: integer
      vehicle_warranty:
        $ref: '#/definitions/VehicleWarranty'
    type: object
  VehicleModel:
    properties:
      description:
        type: string
      name:
        type: string
      vehicle_brand:
        allOf:
        - $ref: '#/definitions/VehicleBrand'
        readOnly: true
    required:
    - description
    - name
    type: object
...
  VehicleModelSchema:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      description:
        type: string
      galleries:
        items:
          $ref: '#/definitions/VehicleGalleryItemSchema'
        type: array
      id:
        format: int32
        type: integer
        x-nullable: true
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
      vehicle_brand:
        allOf:
        - $ref: '#/definitions/VehicleBrand'
        readOnly: true
      vehicle_brand_id:
        format: int32
        type: integer
    required:
    - description
    - name
    type: object
...
  VehicleModelStyleSchema:
    properties:
      base_MSRP:
        format: float
        type: number
      body:
        allOf:
        - $ref: '#/definitions/VehicleBody'
        readOnly: true
      body_id:
        format: int32
        type: integer
      consumer_info:
        type: object
      creation_date:
        format: date-time
        readOnly: true
        type: string
      current_price_assigment:
        items:
          $ref: '#/definitions/PriceVehicleModelStyleAssignmentSchema'
        type: array
      description:
        type: string
      drive_trains:
        items:
          $ref: '#/definitions/VehicleDriveTrain'
        type: array
      engines:
        items:
          $ref: '#/definitions/VehicleEngine'
        type: array
      galleries:
        items:
          $ref: '#/definitions/VehicleGalleryItemSchema'
        type: array
      id:
        format: int32
        type: integer
        x-nullable: true
      image_url:
        type: string
      include_feature_description:
        type: string
      infos:
        items:
          $ref: '#/definitions/VehicleInfo'
        type: array
      manufacture_code:
        type: string
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/VehicleOptionItem'
        type: array
      pricing_options:
        type: object
      standard_feature:
        type: object
      style_features:
        items:
          type: string
        type: array
      technical_feature:
        type: object
      transmissions:
        items:
          $ref: '#/definitions/VehicleTransmission'
        type: array
      vehicle_year_model:
        allOf:
        - $ref: '#/definitions/VehicleYearModel'
        readOnly: true
      vehicle_year_model_id:
        format: int32
        type: integer
      vif_num:
        format: int32
        readOnly: true
        type: integer
    required:
    - description
    - name
    type: object
...
  VehicleOptionItem:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      current_price_assigment:
        items:
          $ref: '#/definitions/PriceOptionAssignmentSchema'
        type: array
      description:
        type: string
      detail:
        type: object
      galleries:
        items:
          $ref: '#/definitions/VehicleGalleryItemSchema'
        type: array
      id:
        format: int32
        type: integer
        x-nullable: true
      is_popular:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      type:
        enum:
        - INTERIOR_COLOR
        - EXTERIOR_COLOR
        - PACKAGE
        - OPTION
        - OTHERS
        type: string
    required:
    - description
    - type
    type: object
...
  VehicleTransmission:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      label:
        type: string
      modification_date:
        format: date-time
        readOnly: true
        type: string
      speeds:
        format: int32
        type: integer
      type:
        enum:
        - OTHERS
        - AUTOMATIC
        - MANUAL
        type: string
    required:
    - type
    type: object
  VehicleWarranty:
    properties:
      anti_corrosion:
        type: string
      comprehensive:
        type: string
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      paint:
        type: string
      power_train:
        type: string
      roadside_assistance:
        type: string
    type: object
  VehicleYearModel:
    properties:
      description:
        type: string
      name:
        type: string
      vehicle_model:
        allOf:
        - $ref: '#/definitions/VehicleModel'
        readOnly: true
    required:
    - description
    - name
    type: object
...
info:
  title: Swagger Petstore
  version: 1.0.0
paths:
...
  /vehicle_attributes:
    get:
      parameters:
      - in: query
        name: direction
        required: 'true'
        schema:
          default: asc
          enum:
          - asc
          - desc
          type: string
      - in: query
        name: is_active
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      - in: query
        name: limit
        required: 'true'
        schema:
          default: '100'
          type: integer
      - in: query
        name: offset
        required: 'true'
        schema:
          default: '0'
          type: integer
      - default: ''
        in: query
        name: q
        required: 'true'
        schema:
          type: string
      - in: query
        name: sort
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema: VehicleAttributeSchema
        '400': &id010
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401': &id011
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX: &id012
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    post:
      parameters:
      - in: body
        name: VehicleAttributeBody
        required: 'true'
        schema:
          $ref: '#/definitions/VehicleAttributeSchema'
      responses:
        '200':
          content:
            application/json:
              schema: VehicleAttributeSchema
        '400': *id010
        '401': *id011
        5XX: *id012
  /vehicle_attributes/{vehicle_attribute_id}:
    delete:
      responses:
        '200':
          content:
            application/json:
              schema: VehicleAttributeSchema
        '400': &id013
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401': &id014
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX: &id015
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    get:
      responses:
        '200':
          content:
            application/json:
              schema: VehicleAttributeSchema
        '400': *id013
        '401': *id014
        5XX: *id015
    parameters:
    - description: VehicleAttributeId
      in: path
      name: vehicle_attribute_id
      required: true
    put:
      parameters:
      - in: body
        name: VehicleAttributeBody
        required: 'true'
        schema:
          $ref: '#/definitions/VehicleAttributeSchema'
      responses:
        '200':
          content:
            application/json:
              schema: VehicleAttributeSchema
        '400': *id013
        '401': *id014
        5XX: *id015
...
  /vehicle_brands/{vehicle_brand_id}/vehicle_models:
    get:
      parameters:
      - in: query
        name: direction
        required: 'true'
        schema:
          default: asc
          enum:
          - asc
          - desc
          type: string
      - in: query
        name: is_active
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      - in: query
        name: limit
        required: 'true'
        schema:
          default: '100'
          type: integer
      - in: query
        name: offset
        required: 'true'
        schema:
          default: '0'
          type: integer
      - default: ''
        in: query
        name: q
        required: 'true'
        schema:
          type: string
      - in: query
        name: sort
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema: VehicleModelSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    parameters:
    - description: VehicleBrandId
      in: path
      name: vehicle_brand_id
      required: true
  /vehicle_model_styles:
    get:
      parameters:
      - in: query
        name: direction
        required: 'true'
        schema:
          default: asc
          enum:
          - asc
          - desc
          type: string
      - in: query
        name: is_active
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      - in: query
        name: limit
        required: 'true'
        schema:
          default: '100'
          type: integer
      - in: query
        name: offset
        required: 'true'
        schema:
          default: '0'
          type: integer
      - default: ''
        in: query
        name: q
        required: 'true'
        schema:
          type: string
      - in: query
        name: sort
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema: VehicleModelStyleSchema
        '400': &id022
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401': &id023
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX: &id024
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    post:
      parameters:
      - in: body
        name: VehicleModelStyleBody
        required: 'true'
        schema:
          $ref: '#/definitions/VehicleModelStyleSchema'
      responses:
        '200':
          content:
            application/json:
              schema: VehicleModelStyleSchema
        '400': *id022
        '401': *id023
        5XX: *id024
...
  /vehicle_model_styles/{vehicle_model_style_id}:
    delete:
      responses:
        '200':
          content:
            application/json:
              schema: VehicleModelStyleSchema
        '400': &id025
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401': &id026
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX: &id027
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    get:
      responses:
        '200':
          content:
            application/json:
              schema: VehicleModelStyleSchema
        '400': *id025
        '401': *id026
        5XX: *id027
    parameters:
    - description: vehicleModelStyleId
      in: path
      name: vehicle_model_style_id
      required: true
    put:
      parameters:
      - in: body
        name: VehicleModelStyleBody
        required: 'true'
        schema:
          $ref: '#/definitions/VehicleModelStyleSchema'
      responses:
        '200':
          content:
            application/json:
              schema: VehicleModelStyleSchema
        '400': *id025
        '401': *id026
        5XX: *id027
...
  /vehicle_models:
    get:
      parameters:
      - in: query
        name: direction
        required: 'true'
        schema:
          default: asc
          enum:
          - asc
          - desc
          type: string
      - in: query
        name: is_active
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      - in: query
        name: limit
        required: 'true'
        schema:
          default: '100'
          type: integer
      - in: query
        name: offset
        required: 'true'
        schema:
          default: '0'
          type: integer
      - default: ''
        in: query
        name: q
        required: 'true'
        schema:
          type: string
      - in: query
        name: sort
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema: VehicleModelSchema
        '400': &id028
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401': &id029
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX: &id030
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    post:
      parameters:
      - in: body
        name: VehicleModelBody
        required: 'true'
        schema:
          $ref: '#/definitions/VehicleModelSchema'
      responses:
        '200':
          content:
            application/json:
              schema: VehicleModelSchema
        '400': *id028
        '401': *id029
        5XX: *id030
  /vehicle_models/{vehicle_model_id}:
    delete:
      responses:
        '200':
          content:
            application/json:
              schema: VehicleModelSchema
        '400': &id031
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401': &id032
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX: &id033
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    get:
      responses:
        '200':
          content:
            application/json:
              schema: VehicleModelSchema
        '400': *id031
        '401': *id032
        5XX: *id033
    parameters:
    - description: vehicleModelId
      in: path
      name: vehicle_model_id
      required: true
    put:
      parameters:
      - in: body
        name: VehicleModelBody
        required: 'true'
        schema:
          $ref: '#/definitions/VehicleModelSchema'
      responses:
        '200':
          content:
            application/json:
              schema: VehicleModelSchema
        '400': *id031
        '401': *id032
        5XX: *id033
...
  /vehicle_year_models/{vehicle_year_model_id}/vehicle_model_styles:
    get:
      parameters:
      - in: query
        name: direction
        required: 'true'
        schema:
          default: asc
          enum:
          - asc
          - desc
          type: string
      - in: query
        name: is_active
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      - in: query
        name: limit
        required: 'true'
        schema:
          default: '100'
          type: integer
      - in: query
        name: offset
        required: 'true'
        schema:
          default: '0'
          type: integer
      - default: ''
        in: query
        name: q
        required: 'true'
        schema:
          type: string
      - in: query
        name: sort
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema: VehicleModelStyleSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    parameters:
    - description: vehicleYearModelId
      in: path
      name: vehicle_year_model_id
      required: true
  /vehicles/_search:
    get:
      parameters:
      - in: query
        name: filter_search
        required: false
        type: string
      - format: int32
        in: query
        name: filter_offset
        required: true
        type: integer
      - format: int32
        in: query
        name: filter_limit
        required: true
        type: integer
      - format: float
        in: query
        name: filter_price_to
        required: false
        type: number
      - format: int32
        in: query
        name: filter_model_id
        required: false
        type: integer
      - in: query
        name: filter_model_style_color_code
        required: false
        type: string
      - format: int32
        in: query
        name: filter_brand_id
        required: false
        type: integer
      - format: int32
        in: query
        name: filter_year_model_id
        required: false
        type: integer
      - format: int32
        in: query
        name: filter_body_id
        required: false
        type: integer
      - format: int32
        in: query
        name: filter_model_style_id
        required: false
        type: integer
      - format: float
        in: query
        name: filter_price_from
        required: false
        type: number
      responses:
        '200':
          content:
            application/json:
              schema: VehicleSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
  /vehicles/{vehicle_id}:
    get:
      parameters:
      - description: VehicleId
        in: path
        name: vehicle_id
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: VehicleSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
...
swagger: 2.0.0